fix: add WASI impls to linker
This commit adds the missing WASI impls that are required by the
golang and python components to the linker.

While the Rust component that is generated is very svelte (no
extraneous WASI imports like `wasi:cli/stdout`), the Golang and Python
WebAssembly components have some WASI imports remaining that must be
filled (even if they are not necessarily used), before execution.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>